==> 02c_combine_generated.ipynb <==
datapath = "./dataset/" #path to the dataset folder

import pandas as pd
import regex
from polyglot.text import Text, Word
from tqdm import tqdm

multisocial = pd.read_csv(datapath + f'multisocial.csv.gz', lineterminator='\n')

models = ['vicuna-13b', 'aya-101', 'v5-Eagle-7B-HF', 'Mistral-7B-Instruct-v0.2', 'opt-iml-max-30b', 'gpt-3.5-turbo-0125', 'gemini']#, 'Llama-2-70b-chat-hf']
datasets = {}

for model in models:
  temp = pd.read_csv(datapath + f'multisocial_{model}_paraphrased_{model}_paraphrased_{model}.csv.gz', lineterminator='\n')
  temp.fillna("", inplace=True)
  temp['text'] = multisocial['text']
  temp.to_csv(f'multisocial_3x_{model}.csv.gz')
  datasets[model] = temp

#check for any duplicates in human data
multisocial[multisocial.text.duplicated(keep=False)]

#remove whitespaces around texts
def clear_dataset(df):
  df_string_columns = df.select_dtypes(['object'])
  df[df_string_columns.columns] = df_string_columns.apply(lambda x: x.str.strip())
  return df

#remove some unicode chars making problems in polyglot
#https://github.com/aboSamoor/polyglot/issues/71#issuecomment-707997790
def remove_bad_chars(text):
  RE_BAD_CHARS = regex.compile(r"[\p{Cc}\p{Cs}]+")
  return RE_BAD_CHARS.sub("", text)

#shorten generated texts
def shorten_text(row, index):
  human = str(row.generated).strip()
  if (human == ''):
    return human
  human_words = Text(human, hint_language_code=row.language).words
  human_length = len(human_words)

  while ((human_length > 200)): #remove last word if text longer than 200 words
    human = human[:-len(str(human_words[-1]))].strip()
    human_words = human_words[:-1]
    human_length = len(human_words)
  return human

#unify dataset form
def unify_form(dataset, model):
  dataset = clear_dataset(dataset)
  dataset['label'] = model
  dataset['text'] = dataset['generated']
  dataset['length'] = [len(x.split()) if (y != 'zh') or (x == '') else len(Text(x, hint_language_code=y).words) for (x, y) in zip(dataset.text, dataset.language)]
  dataset['source'] = [f'multisocial_{x}' for x in dataset.source]
  return dataset

#uniqueness/repetitiveness - get number of unique sentences in row.text
def unique_sentences(row):
  if row.text == '':
    return 0
  sentences = Text(row.text, hint_language_code=row.language).sentences
  return len(set(sentences)) / len(sentences)

#uniqueness/repetitiveness - get number of unique words in row.text
def unique_words(row):
  if row.text == '':
    return 0
  words = Text(row.text, hint_language_code=row.language).words
  return len(set(words)) / len(words)

%%time
stat = {}
for model, dataset in datasets.items():
  print(f'Processing {model}')

  dataset['generated'] = dataset['generated'].apply(lambda x: remove_bad_chars(x))
  dataset['generated'] = [shorten_text(row, index) for index, row in tqdm(dataset.iterrows(), total=len(dataset))]
  empty_generation = len(dataset[(dataset.generated == '') | (dataset.generated == 'nan') | dataset.generated.isna()])

  dataset = unify_form(dataset, model)

  shorts = len(dataset[dataset.length < 2])
  duplicates = len(dataset[dataset.text.duplicated(keep=False)])

  dataset['unique_sentences'] = [unique_sentences(row) for index, row in tqdm(dataset.iterrows())]
  dataset['unique_words'] = [unique_words(row) for index, row in tqdm(dataset.iterrows())]

  stat[model] = {'empty_generation': empty_generation, 'short_texts': shorts, 'duplicates': duplicates, 'wordcount_mean': dataset.length.mean(), 'wordcount_std': dataset.length.std(), 'unique_sentences_mean': dataset.unique_sentences.mean(), 'unique_sentences_std': dataset.unique_sentences.std(), 'unique_words_mean': dataset.unique_words.mean(), 'unique_words_std': dataset.unique_words.std()}
  multisocial = pd.concat([multisocial, dataset], ignore_index=True, copy=False)

multisocial.label.value_counts()

#are instruction-based prompts removed? it's ok
multisocial[multisocial.text.str.contains('You are a helpful assistent.')].label.value_counts()

multisocial[multisocial.text.str.lower().str.contains('assistent')].label.value_counts()

multisocial[multisocial.text.str.lower().str.contains('instruction')].label.value_counts()

multisocial[multisocial.text.str.lower().str.contains('task')].label.value_counts()

multisocial[multisocial.text.str.lower().str.contains('social media text')].label.value_counts()

pd.options.display.float_format = "{:,.2f}".format
pd.DataFrame(stat).T

multisocial[~multisocial.label.str.contains('human')].length.describe()

multisocial[multisocial.label.str.contains('human')].length.describe()

multisocial[multisocial.label.str.contains('human')].length.max()

#shorten human texts
def shorten_text(row, index):
  human = str(row.text).strip()
  if (human == ''):
    return human
  human_words = Text(human, hint_language_code=row.language).words
  human_length = len(human_words)

  while ((human_length > 200)): #remove last word if text longer than 200 words
    human = human[:-len(str(human_words[-1]))].strip()
    human_words = human_words[:-1]
    human_length = len(human_words)
  return human

row = multisocial.iloc[46025]
human = str(row.text).strip()
Text(human, hint_language_code=row.language).words
print(row.text)
print('*****************SHORTENED******************\n', shorten_text(row, 0))

#fix human texts

dataset = multisocial[multisocial.label.str.contains('human')].copy()
dataset['text'] = dataset['text'].apply(lambda x: remove_bad_chars(x).strip())
dataset['text'] = [shorten_text(row, index) for index, row in tqdm(dataset.iterrows(), total=len(dataset))]
dataset['length'] = [len(x.split()) if (y != 'zh') or (x == '') else len(Text(x, hint_language_code=y).words) for (x, y) in zip(dataset.text, dataset.language)]
dataset['unique_sentences'] = [unique_sentences(row) for index, row in tqdm(dataset.iterrows(), total=len(dataset))]
dataset['unique_words'] = [unique_words(row) for index, row in tqdm(dataset.iterrows(), total=len(dataset))]
multisocial.loc[multisocial.label.str.contains('human'),:] = dataset

empty_generation = len(dataset[(dataset.text == '') | (dataset.text == 'nan') | dataset.text.isna()])
shorts = len(dataset[dataset.length < 3])
duplicates = len(dataset[dataset.text.duplicated(keep=False)])

stat['human'] = {'empty_generation': empty_generation, 'short_texts': shorts, 'duplicates': duplicates, 'wordcount_mean': dataset.length.mean(), 'wordcount_std': dataset.length.std(), 'unique_sentences_mean': dataset.unique_sentences.mean(), 'unique_sentences_std': dataset.unique_sentences.std(), 'unique_words_mean': dataset.unique_words.mean(), 'unique_words_std': dataset.unique_words.std()}

pd.DataFrame(stat).T

temp = pd.DataFrame(stat).T.reset_index()
print(temp.to_latex(index=False, na_rep=0, escape=False, formatters={"text": str.lower}, float_format="{:.2f}".format))

multisocial[multisocial.label.str.contains('human')].length.describe()

multisocial[~multisocial.label.str.contains('human')].length.describe()

multisocial.head()

multisocial.drop(columns=['Unnamed: 0'], inplace=True)

multisocial.label.value_counts()

#delete empty and too-short (less than 3 words) texts
multisocial.drop(columns=['generated', 'unique_sentences',	'unique_words'], inplace=True)
multisocial.loc[multisocial.text == "nan", "text"] = pd.NA
multisocial.loc[multisocial.text == "", "text"] = pd.NA
multisocial.dropna(inplace=True)
multisocial = multisocial[multisocial.length > 2]

multisocial.label.value_counts()

#delete text duplicates
multisocial = multisocial.drop_duplicates(subset=['text'])

multisocial.label.value_counts()

multisocial.language.value_counts()

multisocial.groupby(['label'])['language'].value_counts()

multisocial.split.value_counts()

multisocial.groupby(['split'])['language'].value_counts()

multisocial['multi_label'] = multisocial['label'].copy()
multisocial['label'] = int(0)
multisocial.loc[~multisocial.multi_label.str.contains('human'), 'label'] = int(1)
multisocial = multisocial.sample(frac=1., random_state = 0).reset_index(drop=True)
multisocial.to_csv(datapath + f'multisocial_v1.csv.gz', index=False)

multisocial.head()

print(multisocial[multisocial.multi_label.str.contains('human')].length.describe())
print(multisocial[~multisocial.multi_label.str.contains('human')].length.describe())
